pyproject.toml
[project]
name = "playkg"
version = "0.1.0"
description = "Knowledge graph playground with Neo4j, LangChain, and Ollama integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "python-dotenv",
  "PyYAML",
  "termcolor",
  "neo4j",
  "langchain-neo4j",
  "ollama",
  "requests",
  "langchain-community",
  "langchain-ollama",
  "langchain-google-vertexai",
  "folium"
]